Restructure modules.
